[msbuild] Merge the CollectITunesArtwork and CollectITunesArtworkTaskBase classes.
We no longer need two have overridable logic for remote builds, so the
non-abstract task class and the abstract base class can be merged.
