--- msbuild/Xamarin.iOS.Tasks/Tasks/CollectITunesArtwork.cs
@@ -1,28 +1,0 @@
-using System.Collections.Generic;
-using System.Linq;
-using Microsoft.Build.Framework;
-using Xamarin.Messaging.Build.Client;
-
-namespace Xamarin.iOS.Tasks {
-	public class CollectITunesArtwork : CollectITunesArtworkTaskBase, ITaskCallback, ICancelableTask {
-		public override bool Execute ()
-		{
-			if (ShouldExecuteRemotely ())
-				return new TaskRunner (SessionId, BuildEngine4).RunAsync (this).Result;
-
-			return base.Execute ();
-		}
-
-		public bool ShouldCopyToBuildServer (ITaskItem item) => true;
-
-		public bool ShouldCreateOutputFile (ITaskItem item) => false;
-
-		public IEnumerable<ITaskItem> GetAdditionalItemsToBeCopied () => Enumerable.Empty<ITaskItem> ();
-
-		public void Cancel ()
-		{
-			if (ShouldExecuteRemotely ())
-				BuildConnection.CancelAsync (BuildEngine4).Wait ();
-		}
-	}
-}
